+++ src/tightdb/link_view.cpp
@@ -160,39 +160,6 @@ void LinkView::clear()
 #endif
 }
 
-
-namespace {
-
-template<class T> struct LinkComparer {
-    LinkComparer(size_t column, bool ascend, LinkView& lv):
-        m_column(column),
-        m_ascending(ascend),
-        m_lv(lv)
-    {
-    }
-
-    bool operator()(size_t i, size_t j) const
-    {
-        T v1 = m_lv.get_value<T>(i, m_column);
-        T v2 = m_lv.get_value<T>(j, m_column);
-        bool b = CompareLess<T>::compare(v1, v2);
-        return m_ascending ? b : !b;
-    }
-
-    size_t m_column;
-    bool m_ascending;
-    LinkView& m_lv;
-};
-
-} // anonymous namespace
-
-/*
-void LinkView::sort(size_t column_ndx, bool ascending)
-{
-    sort(column_ndx, ascending);
-}
-*/
-
 TableView LinkView::get_sorted_view(size_t column_ndx, bool ascending)
 {
     TableView v(m_origin_column.get_target_table());
@@ -202,51 +169,6 @@ TableView LinkView::get_sorted_view(size_t column_ndx, bool ascending)
     return v;
 }
 
-/*
-template <class T> void LinkView::sort(size_t column_ndx, Column& dest, bool ascending)
-{
-    vector<size_t> v, v2;
-    for (size_t t = 0; t < size(); t++) {
-        v.push_back(t);
-        v2.push_back(m_row_indexes.get(t));
-    }
-    LinkComparer<T> c = LinkComparer<T>(column_ndx, ascending, *this);
-    stable_sort(v2.begin(), v2.end(), c);
-    dest.clear();
-    for (size_t t = 0; t < v.size(); t++)
-        dest.add(v2[t]);
-}
-
-
-void LinkView::sort(size_t column_ndx, Column& dest, bool ascending)
-{
-    Table& target_table = m_origin_column.get_target_table();
-    DataType type = target_table.get_column_type(column_ndx);
-
-    TIGHTDB_ASSERT(type == type_Int ||
-        type == type_DateTime ||
-        type == type_Bool ||
-        type == type_Float ||
-        type == type_Double ||
-        type == type_String);
-
-    if (m_row_indexes.size() == 0)
-        return;
-
-    if (type == type_Float) {
-        sort<float>(column_ndx, dest, ascending);
-    }
-    else if (type == type_Double) {
-        sort<double>(column_ndx, dest, ascending);
-    }
-    else if (type == type_String) {
-        sort<StringData>(column_ndx, dest, ascending);
-    }
-    else {
-        sort<int64_t>(column_ndx, dest, ascending);
-    }
-}
-*/
 
 void LinkView::remove_target_row(size_t link_ndx)
 {