+++ src/tightdb/row.cpp
@@ -5,22 +5,22 @@ using namespace std;
 using namespace tightdb;
 
 
-void RowBase::attach(Table* table, size_t row_ndx)
+void RowBase::attach(Table* table, size_t row_ndx) TIGHTDB_NOEXCEPT
 {
     if (table) {
-        table->register_row_accessor(this); // Throws
+        table->register_row_accessor(this);
         m_table.reset(table);
         m_row_ndx = row_ndx;
     }
 }
 
-void RowBase::reattach(Table* table, size_t row_ndx)
+void RowBase::reattach(Table* table, size_t row_ndx) TIGHTDB_NOEXCEPT
 {
     if (m_table.get() != table) {
-        if (table)
-            table->register_row_accessor(this); // Throws
         if (m_table)
             m_table->unregister_row_accessor(this);
+        if (table)
+            table->register_row_accessor(this);
         m_table.reset(table);
     }
     m_row_ndx = row_ndx;