+++ release_notes.md
@@ -20,6 +20,15 @@
 
 ----------------------------------------------
 
+# 0.82.2 Release notes
+
+### Internals:
+
+* Query must now be deep-copied using the '=' operator instead of using
+  TCopyExpressionTag. Also fixed a bug in this deep-copying.
+
+----------------------------------------------
+
 # 0.82.1 Release notes
 
 ### Internals: